backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0
asyncpg>=0.29.0
alembic>=1.13.1
# pgvector>=0.2.4  # Uncomment for PostgreSQL with vector search

# Redis & Task Queue
redis>=5.0.1
arq>=0.25.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

# Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.1

# Document Processing
pymupdf>=1.23.8
python-docx>=1.1.0
rapidocr-onnxruntime>=1.3.8

# Document Generation
python-docx>=1.1.0
weasyprint>=60.2

# Data Processing
pyyaml>=6.0.1
numpy>=1.26.3

# Web Scraping (for RAG)
beautifulsoup4>=4.12.3
lxml>=5.1.0

# Security
cryptography>=42.0.0

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.3
loguru>=0.7.2

# OpenTelemetry (observability)
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-exporter-otlp-proto-grpc>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-instrumentation-httpx>=0.44b0
opentelemetry-instrumentation-sqlalchemy>=0.44b0
opentelemetry-instrumentation-redis>=0.44b0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
